change in where to broadcast
removed broadcast from handleLocalRequest(), as the order is sent on requestToNetwork channel. This broadcast did also cause the program to crash when running multiple instances.
Added broadcast when we delegate order to the local fsm.

Changed delegation timeout, I experienced several deadlocks with the previous 50 ms when running 3 elevators. This should be investigated further.